route for test flight data view
