+++ frontend/package.json
@@ -10,9 +10,11 @@
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
+    "bootstrap": "^5.3.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-media-recorder": "^1.6.6",
+    "react-router-dom": "^6.16.0",
     "react-scripts": "^5.0.0",
     "styled-components": "^5.3.5",
     "web-vitals": "^2.1.3",
