+++ frontend/src/components/Header.js
@@ -3,7 +3,7 @@ import { Box, AppBar, Typography } from '@mui/material';
 
 function Header() {
     return(
-        <Box sx={{ flexGrow: 1, backgroundColor: '#777772' }}>
+        <Box sx={{ flexGrow: 1, backgroundColor: '#777772', marginBottom: '1vh' }}>
             <AppBar position="static">
                 <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
                     Western Aero Design Team - Ground Station
